fixed based on review (#31)
