+++ src/main/resources/META-INF/plugin.xml
@@ -36,7 +36,7 @@
       <action class="com.redhat.devtools.intellij.tektoncd.actions.resource.CreateResourceAction" text="New Pipeline Resource"/>
       <action class="com.redhat.devtools.intellij.tektoncd.actions.OpenEditorTaskAction" text="Open in Editor"/>
       <action class="com.redhat.devtools.intellij.tektoncd.actions.task.StartTaskAction" text="Start"/>
-      <action class="com.redhat.devtools.intellij.tektoncd.actions.component.ShowLogsAction" text="Show Logs"/>
+      <action class="com.redhat.devtools.intellij.tektoncd.actions.ShowLogsAction" text="Show Logs"/>
       <action class="com.redhat.devtools.intellij.tektoncd.actions.DeleteAction" text="Delete"/>
       <action class="com.redhat.devtools.intellij.common.actions.RefreshAction" text="Refresh"/>
     </group>
